Store subcube matrices in main storage bucket
