+++ db.js
@@ -4,7 +4,7 @@ export async function openDB() {
     if (dbOpenPromise) return dbOpenPromise;
     dbOpenPromise = new Promise((resolve, reject) => {
         const openIndexedDB = (idbFactory) => {
-            const request = idbFactory.open('CubeDB', 2);
+            const request = idbFactory.open('CubeDB', 3);
             request.onupgradeneeded = (event) => {
                 const db = event.target.result;
                 if (!db.objectStoreNames.contains('cubes')) {
@@ -21,6 +21,9 @@ export async function openDB() {
                     vertStore.createIndex('subCubeId', 'subCubeId', { unique: false });
                     vertStore.createIndex('cubeId', 'cubeId', { unique: false });
                     vertStore.createIndex('windowUID', 'windowUID', { unique: false });
+                }
+                if (!db.objectStoreNames.contains('matrices')) {
+                    db.createObjectStore('matrices', { keyPath: 'id' });
                 }
             };
             request.onsuccess = () => resolve(request.result);
@@ -249,4 +252,24 @@ export async function cleanupStaleWindows(db, validIds) {
     for (let id of [...new Set(toDelete)]) {
         await deleteWindowData(db, id);
     }
+}
+
+export async function saveMatrix(db, cubeId, subId, matrix) {
+    return new Promise((resolve, reject) => {
+        const tx = db.transaction('matrices', 'readwrite');
+        const store = tx.objectStore('matrices');
+        store.put({ id: `${cubeId}_${subId}`, cubeId, subId, matrix });
+        tx.oncomplete = () => resolve();
+        tx.onerror = () => reject(tx.error);
+    });
+}
+
+export async function loadMatrix(db, cubeId, subId) {
+    return new Promise((resolve, reject) => {
+        const tx = db.transaction('matrices', 'readonly');
+        const store = tx.objectStore('matrices');
+        const req = store.get(`${cubeId}_${subId}`);
+        req.onsuccess = () => resolve(req.result ? req.result.matrix : null);
+        req.onerror = () => reject(req.error);
+    });
 }
